+++ screens/ProfileScreen.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useMemo, useEffect } from "react";
 import {
    View,
    Text,
@@ -11,17 +11,34 @@ import {
 import { useSelector, useDispatch } from "react-redux";
 
 import { startLogout } from "../store/actions/auth";
+import { setLanguage } from "../store/actions/lang";
 
 import labels from "../constants/labels";
 import Colors from "../constants/colors";
+import InputField from "../components/InputField";
+import { formatValue } from "../utils";
 
 const ProfileScreen = (props) => {
    const user = useSelector((state) => state.auth);
+   const expenses = useSelector((state) => state.expenses);
    const dispatch = useDispatch();
+
+   const locale = useSelector((state) => state.lang.locale);
+   const dictonary = labels[locale].ProfileScreen;
+
+   const expenseTotal = useMemo(() => {
+      return expenses.reduce((sum, expense) => sum + expense.amount, 0);
+   }, [expenses]);
 
    const signOutHandler = () => {
       dispatch(startLogout());
    };
+
+   useEffect(() => {
+      props.navigation.setOptions({
+         headerTitle: dictonary.title,
+      });
+   }, [locale]);
 
    return (
       <View style={styles}>
@@ -37,14 +54,47 @@ const ProfileScreen = (props) => {
 
                <TouchableOpacity activeOpacity={0.6} onPress={signOutHandler}>
                   <View style={styles.button}>
-                     <Text style={styles.buttonText}>
-                        {labels.es.ProfileScreen.signOut}
-                     </Text>
+                     <Text style={styles.buttonText}>{dictonary.signOut}</Text>
                   </View>
                </TouchableOpacity>
             </View>
          </View>
-         <View></View>
+         <View style={styles.boxContainer}>
+            <View style={styles.box}>
+               <Text style={styles.boxTitle}>
+                  {formatValue("currency", expenseTotal / 100)}
+               </Text>
+               <Text style={styles.boxContent}>
+                  {dictonary.expenseTotalMessage}
+               </Text>
+            </View>
+            <View style={styles.box}>
+               <Text style={styles.boxTitle}>{expenses.length}</Text>
+               <Text style={styles.boxContent}>
+                  {dictonary.expenseCountMessage}
+               </Text>
+            </View>
+         </View>
+         <View>
+            <InputField
+               label={dictonary.inputText}
+               type="select"
+               options={[
+                  {
+                     label: "Español",
+                     value: "es",
+                  },
+                  {
+                     label: "English",
+                     value: "en",
+                  },
+               ]}
+               value={locale}
+               onChangeText={(value) => {
+                  dispatch(setLanguage(value));
+               }}
+            />
+         </View>
       </View>
    );
 };
@@ -66,7 +116,7 @@ const styles = StyleSheet.create({
    },
    userInfoContainer: {
       flexDirection: "row",
-      backgroundColor: Colors.gray,
+      backgroundColor: Colors.darkGray,
       padding: 30,
    },
    usernameContainer: {
@@ -80,6 +130,27 @@ const styles = StyleSheet.create({
       color: Colors.primary,
       fontSize: 20,
       // fontWeight: "bold",
+   },
+   boxContainer: {
+      flexDirection: "row",
+   },
+   box: {
+      borderWidth: 1,
+      borderColor: Colors.borderColor,
+      flex: 1,
+      padding: 20,
+   },
+   boxTitle: {
+      fontFamily: "roboto-bold",
+      color: Colors.primary,
+      fontSize: 18,
+      textAlign: "center",
+   },
+   boxContent: {
+      fontFamily: "roboto",
+      color: Colors.grayBorder,
+      fontWeight: "bold",
+      textAlign: "center",
    },
 });