+++ screens/ReportMonthScreen.js
@@ -1,0 +1,242 @@
+import React, { useState, useEffect, useMemo, useCallback } from "react";
+import {
+   View,
+   Text,
+   StyleSheet,
+   TouchableOpacity,
+   ScrollView,
+   Dimensions,
+} from "react-native";
+import { useSelector } from "react-redux";
+
+import labels from "../constants/labels";
+import Colors from "../constants/colors";
+import navigation from "../constants/navigation";
+import Chart from "../components/Chart";
+
+import moment from "moment";
+import capitalize from "lodash/capitalize";
+import times from "lodash/times";
+import { replaceAll, formatValue } from "../utils";
+
+import { useHeaderHeight } from "@react-navigation/stack";
+
+const getDateAsText = (date) => {
+   const text = date.format("MMMM");
+   return capitalize(replaceAll(text, "\\.", ""));
+};
+
+const ReportMonthScreen = (props) => {
+   const [today, setToday] = useState(moment().startOf("month"));
+   const expenses = useSelector((state) => state.expenses);
+
+   const sixMonthAgo = moment(today).subtract(5, "month");
+
+   const sixMonthAgoLabels = times(6, (n) => {
+      return moment(sixMonthAgo).add(n, "month").format("MMM YY");
+   });
+
+   const filterByDate = (expense) => {
+      let isValid = true;
+      if (
+         moment(expense.createdAt).isBefore(sixMonthAgo.startOf("month")) ||
+         moment(expense.createdAt).isAfter(today.endOf("month"))
+      ) {
+         isValid = false;
+      }
+      return isValid;
+   };
+
+   const filterByCategory = (category) => (expense) => {
+      let isValid = true;
+      if (
+         moment(expense.createdAt).isBefore(sixMonthAgo.startOf("month")) ||
+         moment(expense.createdAt).isAfter(today.endOf("month"))
+      ) {
+         isValid = false;
+      }
+      if (expense.category !== category) {
+         isValid = false;
+      }
+      return isValid;
+   };
+
+   const dataFunction = useCallback(
+      (filterFunc) => () => {
+         // Get the expenses that in range.
+         const expensesInRange = expenses.filter(filterFunc);
+
+         const sixMonthAgoObj = times(6, (n) => {
+            return moment(sixMonthAgo).add(n, "month").format("YYYYMM");
+         }).reduce((acum, key) => {
+            return {
+               ...acum,
+               [key]: {
+                  monthString: key,
+               },
+            };
+         }, {});
+
+         // Get the total of each month
+         const rawData = expensesInRange.reduce((acum, expense) => {
+            const month = moment(expense.createdAt).startOf("month");
+            let total = 0;
+            total = expense.amount / 100;
+            const monthString = month.format("YYYYMM");
+            const label = month.format("MMM YY");
+            return {
+               ...acum,
+               [monthString]: acum[monthString]
+                  ? {
+                       ...acum[monthString],
+                       total:
+                          (acum[monthString].total
+                             ? acum[monthString].total
+                             : 0) + total,
+                       label,
+                       monthString,
+                    }
+                  : { total, label, monthString },
+            };
+         }, sixMonthAgoObj);
+
+         // Order data by month
+         return Object.keys(rawData)
+            .map((key) => {
+               const item = rawData[key];
+               return item;
+            })
+            .sort((a, b) => {
+               const aInt = parseInt(a.monthString);
+               const bInt = parseInt(b.monthString);
+               return aInt - bInt;
+            });
+      },
+      [expenses, today]
+   );
+
+   const dataObj = useMemo(dataFunction(filterByDate), [expenses, today]);
+   const foodDataObj = useMemo(dataFunction(filterByCategory("food")), [
+      expenses,
+      today,
+   ]);
+   const billsDataObj = useMemo(dataFunction(filterByCategory("bills")), [
+      expenses,
+      today,
+   ]);
+   const newData = dataObj.map((d) => (d.total ? d.total : 0));
+   const foodData = foodDataObj.map((d) => (d.total ? d.total : 0));
+   const billsData = billsDataObj.map((d) => (d.total ? d.total : 0));
+
+   const line = {
+      labels: sixMonthAgoLabels,
+      datasets: [
+         {
+            data: newData,
+            strokeWidth: 1, // optional
+         },
+      ],
+   };
+   const line2 = {
+      labels: sixMonthAgoLabels,
+      datasets: [
+         {
+            data: foodData,
+            strokeWidth: 1, // optional
+         },
+      ],
+   };
+   const line3 = {
+      labels: sixMonthAgoLabels,
+      datasets: [
+         {
+            data: billsData,
+            strokeWidth: 1, // optional
+         },
+      ],
+   };
+
+   const headerHeight = useHeaderHeight();
+
+   const resultHeight =
+      Dimensions.get("window").height -
+      40 -
+      navigation.BOTTOM_TAB_HEIGHT -
+      headerHeight;
+   return (
+      <View style={styles}>
+         <View style={styles.header}>
+            <Button
+               text="<"
+               onPress={() => {
+                  setToday(moment(today).subtract(1, "month"));
+               }}
+            />
+            <Text>
+               {getDateAsText(sixMonthAgo)} - {getDateAsText(today)}
+            </Text>
+            <Button
+               text=">"
+               onPress={() => {
+                  setToday(moment(today).add(1, "month"));
+               }}
+            />
+         </View>
+
+         <ScrollView
+            style={{
+               height: resultHeight,
+            }}
+         >
+            <Chart
+               title={labels.es.ReportMonthScreen.chartTitle}
+               line={line}
+               data={newData}
+            />
+            <Chart
+               title={labels.es.categories.food}
+               line={line2}
+               data={foodData}
+            />
+            <Chart
+               title={labels.es.categories.bills}
+               line={line3}
+               data={billsData}
+            />
+         </ScrollView>
+      </View>
+   );
+};
+
+const Button = ({ text, onPress }) => {
+   return (
+      <TouchableOpacity onPress={onPress}>
+         <Text style={styles.button}>{text}</Text>
+      </TouchableOpacity>
+   );
+};
+
+const styles = StyleSheet.create({
+   header: {
+      flexDirection: "row",
+      justifyContent: "space-between",
+      alignItems: "center",
+      paddingHorizontal: 40,
+      backgroundColor: Colors.gray,
+      height: 40,
+   },
+   button: {
+      fontSize: 24,
+   },
+   chartContainer: {
+      // height: "90%",
+   },
+});
+
+export const screenOptions = (navData) => {
+   return {
+      headerTitle: labels.es.ReportMonthScreen.title,
+   };
+};
+
+export default ReportMonthScreen;