+++ changelog.md
@@ -1,3 +1,15 @@
+### V0.2
+
+Sept 20, 2020
+
+-  [NEW]
+   -  Expense Filter Screen working
+   -  Reports Screen
+      -  Report Month Screen
+   -  Change language in Profile Screen
+-  [Fix]
+   -  Change from ScrollView to FlatList on ExpenseList component
+
 ### V0.1
 
 Sept 16, 2020